% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequality_indices.R
\name{inequality_indices}
\alias{inequality_indices}
\title{Calcul des indices d'inegalite}
\usage{
inequality_indices(data, variable, separateur = NULL)
}
\arguments{
\item{data}{Un data frame contenant les donnees.}

\item{variable}{Nom de la variable representant le revenu ou la consommation.}

\item{separateur}{Nom de la colonne permettant une analyse par sous-groupes (ex. sexe, region, milieu).}
}
\value{
Une liste contenant :
\itemize{
\item La courbe de Lorenz
\item L'indice de Gini
\item L'indice de Theil
}
}
\description{
Cette fonction permet de calculer plusieurs indices d'inegalite economique,
notamment la courbe de Lorenz, l'indice de Gini et l'indice de Theil.
}
