% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_distribution.R
\name{population_distribution}
\alias{population_distribution}
\title{Analyse la repartition de la population par categories}
\usage{
population_distribution(data, separateur)
}
\arguments{
\item{data}{Un data.frame contenant les donnees.}

\item{separateur}{Une chaîne de caractères indiquant la variable par laquelle effectuer la repartition.}
}
\value{
Un graphique et un tableau recapitulatif de la repartition.
}
\description{
Analyse la repartition de la population par categories
}
